feature: Adds spacer after IntroductionSection.
The ToolSet list that is present in the IntroductionSection would
overlap with the elements that followed it. To avoid this problem a
transparent spacer has been added that immitates the height of the list.
This allows for the list height to remain dynamic(not fixed), while
still not overlapping with subsequent sections.